Add files via upload
This is likely not the easiest way to go about this problem.  I recognized the need to iterate through the variables incrementally, checking each permutation of the multiplication such as, #1 * #1 * #1 then #1 * #1 * #2 then #1 * #1 * #4, etc.  I used a nested set of if statements with ranges i, j, k that are equal to the length of the input data, 200.  The code would hold i and j at value #1 in the input list while incrementing through all 200 values at k.  Once complete, it increments "j" by 1, moving to the second item in the list, while "i" is unaffected and k goes back through all 200 options again, such as: #1 * #2 * #1.  It is not clear to me why the results get printed 6 times.

Requirements below:
--- Part Two ---
The Elves in accounting are thankful for your help; one of them even offers you a starfish coin they had left over from a past vacation. They offer you a second one if you can find three numbers in your expense report that meet the same criteria.

Using the above example again, the three entries that sum to 2020 are 979, 366, and 675. Multiplying them together produces the answer, 241861950.

In your expense report, what is the product of the three entries that sum to 2020?