Add files via upload
completed code to give the correct answer.  Parsing the input data into a list that I could work with took some time.  I wasn't sure at first how to import a txt file and break it down into the list items but various Google searches got me to the a satisfactory result.  Lines 36-39 are admittedly likely not to be needed due to the numbers always incrementing upward from left to right (ie, it will always be 9-11 not 11-9), but at least this captures the corner case.  Also had some issues with getting the value of "good_pw" to increment.  This ended up being due to having the starter variable declaration inside of the for loop instead of globally.